# strutil: make VersionCompare faster