--- a/strutil/version_test.go
+++ b/strutil/version_test.go
@@ -115,8 +115,11 @@ func (s *VersionTestSuite) TestVersionInvalid(c *C) {
 		valid bool
 	}{
 		{"1:2", false},
+		{"12:34", false},
+		{"1234:", false},
 		{"1--1", false},
 		{"1.0", true},
+		{"1234", true},
 	} {
 		res := strutil.VersionIsValid(t.ver)
 		c.Check(res, Equals, t.valid, Commentf("%q: %v but expected %v", t.ver, res, t.valid))