--- a/strutil/version_test.go
+++ b/strutil/version_test.go
@@ -37,6 +37,7 @@ func (s *VersionTestSuite) TestVersionCompare(c *C) {
 		res  int
 		err  error
 	}{
+		{"20000000000000000000", "020000000000000000000", 0, nil},
 		{"1.0", "2.0", -1, nil},
 		{"1.3", "1.2.2.2", 1, nil},
 		{"1.3", "1.3.1", -1, nil},