Per @zyga's request, this splits the `strutil` changes from #6065. It then also adds to them.

With these changes,

    benchmark                   old ns/op     new ns/op     delta
    BenchmarkVersionCompare     19479781      560826        -97.12%

    benchmark                   old allocs     new allocs     delta
    BenchmarkVersionCompare     86629          0              -100.00%

    benchmark                   old bytes     new bytes     delta
    BenchmarkVersionCompare     2976276       0             -100.00%

(it also fixes a bug in VersionCompare that we hadn't hit...)

This is what the cpu profile looked like before:

![old](https://user-images.githubusercontent.com/1543430/47744666-c8093400-dc79-11e8-8ac8-ef8569bcca6d.png)

and this is what it looks like now:

![new](https://user-images.githubusercontent.com/1543430/47744707-e111e500-dc79-11e8-8fc3-71501797efdb.png)